+++ src/main/java/one/talon/model/Application.java
@@ -37,7 +37,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Application {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -271,7 +271,7 @@ public class Application {
   private List<LoyaltyProgram> loyaltyPrograms = new ArrayList<LoyaltyProgram>();
 
 
-  public Application id(Integer id) {
+  public Application id(Long id) {
     
     this.id = id;
     return this;
@@ -283,12 +283,12 @@ public class Application {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
