+++ src/main/java/one/talon/model/ApplicationSession.java
@@ -38,7 +38,7 @@ import org.threeten.bp.OffsetDateTime;
 public class ApplicationSession {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -148,7 +148,7 @@ public class ApplicationSession {
   private Object attributes;
 
 
-  public ApplicationSession id(Integer id) {
+  public ApplicationSession id(Long id) {
     
     this.id = id;
     return this;
@@ -160,12 +160,12 @@ public class ApplicationSession {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
