+++ src/main/java/one/talon/model/Change.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Change {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -63,7 +63,7 @@ public class Change {
   private Integer managementKeyId;
 
 
-  public Change id(Integer id) {
+  public Change id(Long id) {
     
     this.id = id;
     return this;
@@ -75,12 +75,12 @@ public class Change {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
