+++ src/main/java/one/talon/model/CustomEffect.java
@@ -35,7 +35,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CustomEffect {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -90,7 +90,7 @@ public class CustomEffect {
   private Integer createdBy;
 
 
-  public CustomEffect id(Integer id) {
+  public CustomEffect id(Long id) {
     
     this.id = id;
     return this;
@@ -102,12 +102,12 @@ public class CustomEffect {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
