+++ src/main/java/one/talon/model/EventType.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class EventType {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -51,7 +51,7 @@ public class EventType {
   private String description;
 
 
-  public EventType id(Integer id) {
+  public EventType id(Long id) {
     
     this.id = id;
     return this;
@@ -63,12 +63,12 @@ public class EventType {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
