+++ src/main/java/one/talon/model/Export.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Export {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -112,7 +112,7 @@ public class Export {
   private Object filter;
 
 
-  public Export id(Integer id) {
+  public Export id(Long id) {
     
     this.id = id;
     return this;
@@ -124,12 +124,12 @@ public class Export {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
