+++ src/main/java/one/talon/model/GiveawaysPool.java
@@ -35,7 +35,7 @@ import org.threeten.bp.OffsetDateTime;
 public class GiveawaysPool {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -74,7 +74,7 @@ public class GiveawaysPool {
   private Integer modifiedBy;
 
 
-  public GiveawaysPool id(Integer id) {
+  public GiveawaysPool id(Long id) {
     
     this.id = id;
     return this;
@@ -86,12 +86,12 @@ public class GiveawaysPool {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
