+++ src/main/java/one/talon/model/LedgerEntry.java
@@ -33,7 +33,7 @@ import org.threeten.bp.OffsetDateTime;
 public class LedgerEntry {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -72,7 +72,7 @@ public class LedgerEntry {
   private Integer referenceId;
 
 
-  public LedgerEntry id(Integer id) {
+  public LedgerEntry id(Long id) {
     
     this.id = id;
     return this;
@@ -84,12 +84,12 @@ public class LedgerEntry {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
