+++ src/main/java/one/talon/model/ModelReturn.java
@@ -35,7 +35,7 @@ import org.threeten.bp.OffsetDateTime;
 public class ModelReturn {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -78,7 +78,7 @@ public class ModelReturn {
   private Integer createdBy;
 
 
-  public ModelReturn id(Integer id) {
+  public ModelReturn id(Long id) {
     
     this.id = id;
     return this;
@@ -90,12 +90,12 @@ public class ModelReturn {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
