+++ src/main/java/one/talon/model/Store.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Store {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -69,7 +69,7 @@ public class Store {
   private List<Integer> linkedCampaignIds = null;
 
 
-  public Store id(Integer id) {
+  public Store id(Long id) {
     
     this.id = id;
     return this;
@@ -81,12 +81,12 @@ public class Store {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
