+++ src/main/java/one/talon/model/User.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class User {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -150,7 +150,7 @@ public class User {
   private Object additionalAttributes;
 
 
-  public User id(Integer id) {
+  public User id(Long id) {
     
     this.id = id;
     return this;
@@ -162,12 +162,12 @@ public class User {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
